Implement PayloadSender for SwarmTransport
